pyproject.toml
[tool.poetry]
name = "letscountitui"
version = "0.1.0"
description = "A UI for letscountit"
authors = ["Carsten Skov <carsten@simcax.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
gunicorn = "^23.0.0"
requests = "^2.32.3"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
setuptools = "^75.0.0"
ruff = "^0.6.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pre-commit = "^3.8.0"
poetry = "^1.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
